requirements.txt
speechrecognition
pyaudio
edge-tts
google-genai